backend/requirements/base.txt
Django>=5.1.7
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.5.0
drf-yasg>=1.21.7
psycopg2-binary>=2.9.9
redis>=5.0.1
python-dotenv>=1.0.1